Added base for a physics system, rigidbodies, box colliders, sphere colliders, and physics materials.

+++ GraphicsTest/GraphicsTest/InputSystem.cpp
@@ -8,8 +8,8 @@
 
 bool InputSystem::currInputs[NUMINPUTS] = { false };
 bool InputSystem::prevInputs[NUMINPUTS] = { false };
-float InputSystem::xPos = 0;
-float InputSystem::yPos = 0;
+double InputSystem::xPos = 0;
+double InputSystem::yPos = 0;
 
 InputSystem::InputSystem()
 {
